Work around Firestore persistence hating threads
Looks like Firestore REALLY dislikes being called on threads that
aren't the main thread, so for now we avoid subscribing on the IO
scheduler until after we've accessed the ScheduleView. Then, to
avoid at least running the whole pipeline on the main thread, we
do a dirty observeOn(io). It's not great but it gets us 80% of the
way there, at least where we can have it.