Tweak pre-commit hook to not auto-format code before validating it
It was a nice try but it's more pain than it's worth, because any
formatting changes that come up and make it pass the commit will not be
staged and thus won't be committed.

To make it properly work we'd need to parse the git log, stage the
changes, amend the previous commit... that's a lot of effort.